backend/pyproject.toml
[project]
name = "todolistdrf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2>=2.9.10",
    "python-decouple>=3.8",
]
